Try to convert everything to String
Have to use string for automatic mapping to work

+++ TableService/src/main/java/com/restaurant/TableService/core/application/command/AddTableCommand.java
@@ -1,4 +1,4 @@
 package com.restaurant.TableService.core.application.command;
 
-public record AddTableCommand(String id, int numberOfSeats) {
+public record AddTableCommand(String id, long numberOfSeats) {
 }
